Add files via upload
I can't say with certainty that I am convinced about the results, only because K_NN theoretically is sensitive to missing values and refused to respond with the number of missing values that were found in the data sets. I am still of the view, that that is what should have been reported, however at the same time, I understand the viewpoints raised these past couple of days about the handling of missing values for K-NN. Thanks for everyones help throughout. Much appreciated,